+++ src/tests/test_utils.py
@@ -10,7 +10,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 """
-Unit tests for service module
+Unit tests for utils.py
 """
 import json
 import os.path
@@ -33,7 +33,7 @@ sample_path = Path(distribution.location) / "ensembl" / "production" / "metadata
 @pytest.mark.parametrize("multi_dbs", [[{"src": sample_path / "ensembl_metadata"},
                                         {"src": sample_path / "ncbi_taxonomy"}]],
                          indirect=True)
-class TestClass:
+class TestUtils:
     dbc = None  # type: UnitTestDB
 
     @pytest.fixture(scope="class")
@@ -56,144 +56,6 @@ class TestClass:
             metadata_uri=multi_dbs["ensembl_metadata"].dbc.url
         )
         yield release_conn
-
-    def test_create_genome(self, multi_dbs, genome_db_conn):
-        """Test service.create_genome function"""
-        input_data = genome_db_conn.fetch_genomes(genome_uuid="a7335667-93e7-11ec-a39d-005056b38ce3")
-        expected_output = {
-          "genomeUuid": "a7335667-93e7-11ec-a39d-005056b38ce3",
-          "assembly": {
-            "accession": "GCA_000001405.28",
-            "name": "GRCh38.p13",
-            "ucscName": "hg38",
-            "level": "chromosome",
-            "ensemblName": "GRCh38.p13"
-          },
-          "taxon": {
-            "taxonomyId": 9606,
-            "scientificName": "Homo sapiens"
-          },
-          "created": "2023-05-12 13:30:58",
-          "organism": {
-            "displayName": "Human",
-            "scientificName": "Homo sapiens",
-            "urlName": "Homo_sapiens",
-            "ensemblName": "homo_sapiens",
-            "organismUuid": "db2a5f09-2db8-429b-a407-c15a4ca2876d"
-          },
-          "release": {
-            "releaseVersion": 108.0,
-            "releaseDate": "2023-05-15",
-            "releaseLabel": "Beta Release 1",
-            "isCurrent": True,
-            "siteName": "Ensembl",
-            "siteLabel": "Ensembl Genome Browser",
-            "siteUri": "https://beta.ensembl.org"
-          }
-        }
-
-        output = json_format.MessageToJson(utils.create_genome(input_data[0]))
-        assert json.loads(output) == expected_output
-
-    def test_create_assembly(self, multi_dbs, genome_db_conn):
-        input_data = genome_db_conn.fetch_sequences(assembly_uuid="eeaaa2bf-151c-4848-8b85-a05a9993101e")
-        expected_output = {
-            "assemblyUuid": "eeaaa2bf-151c-4848-8b85-a05a9993101e",
-            "accession": "GCA_000001405.28",
-            "level": "chromosome",
-            "name": "GRCh38.p13",
-            "chromosomal": 1,
-            "length": 107043717,
-            "sequenceLocation": "SO:0000738"
-        }
-
-        output = json_format.MessageToJson(utils.create_assembly(input_data[0]))
-        assert json.loads(output) == expected_output
-
-    def test_create_karyotype(self, multi_dbs, genome_db_conn):
-        input_data = genome_db_conn.fetch_sequences(genome_uuid="a7335667-93e7-11ec-a39d-005056b38ce3")
-        expected_output = {
-            "code": "chromosome",
-            "chromosomal": "1",
-            "location": "SO:0000738",
-            "genomeUuid": "a7335667-93e7-11ec-a39d-005056b38ce3"
-        }
-
-        output = json_format.MessageToJson(utils.create_karyotype(input_data[0]))
-        assert json.loads(output) == expected_output
-
-    def test_create_species(self, multi_dbs, genome_db_conn):
-        species_input_data = genome_db_conn.fetch_genomes(genome_uuid="a7335667-93e7-11ec-a39d-005056b38ce3")
-        tax_id = species_input_data[0].Organism.taxonomy_id
-        taxo_results = genome_db_conn.fetch_taxonomy_names(tax_id)
-        expected_output = {
-            "genomeUuid": "a7335667-93e7-11ec-a39d-005056b38ce3",
-            "ncbiCommonName": "human",
-            "taxonId": 9606,
-            "scientificName": "Homo sapiens"
-        }
-
-        output = json_format.MessageToJson(utils.create_species(species_input_data[0], taxo_results[tax_id]))
-        assert json.loads(output) == expected_output
-
-    def test_create_top_level_statistics(self, multi_dbs, genome_db_conn):
-        organism_uuid = "21279e3e-e651-43e1-a6fc-79e390b9e8a8"
-        input_data = genome_db_conn.fetch_genome_datasets(organism_uuid=organism_uuid, dataset_name="all")
-
-        statistics = []
-        # getting just the first element
-        statistics.append({
-            'name': input_data[0].Attribute.name,
-            'label': input_data[0].Attribute.label,
-            'statistic_type': input_data[0].Attribute.type,
-            'statistic_value': input_data[0].DatasetAttribute.value
-        })
-
-        expected_output = {
-            "organismUuid": "21279e3e-e651-43e1-a6fc-79e390b9e8a8",
-            "statistics": [
-                {
-                    "name": "total_genome_length",
-                    "label": "Total genome length",
-                    "statisticType": "bp",
-                    "statisticValue": "4641652"
-                }
-            ]
-        }
-
-        output = json_format.MessageToJson(
-            utils.create_top_level_statistics({
-                'organism_uuid': organism_uuid,
-                'statistics': statistics
-            })
-        )
-        assert json.loads(output) == expected_output
-
-    def test_create_genome_sequence(self, multi_dbs, genome_db_conn):
-        input_data = genome_db_conn.fetch_sequences(genome_uuid="a7335667-93e7-11ec-a39d-005056b38ce3")
-        expected_output = {
-            "accession": "CHR_HG1_PATCH",
-            "name": "CHR_HG1_PATCH",
-            "sequenceLocation": "SO:0000738",
-            "length": 107043717,
-            "chromosomal": True
-        }
-        output = json_format.MessageToJson(utils.create_genome_sequence(input_data[0]))
-        assert json.loads(output) == expected_output
-
-    def test_create_release(self, multi_dbs, release_db_conn):
-        input_data = release_db_conn.fetch_releases(release_version=108)
-        expected_output = {
-            "releaseVersion": 108.0,
-            "releaseDate": "2023-05-15",
-            "releaseLabel": "Beta Release 1",
-            "isCurrent": True,
-            "siteName": "Ensembl",
-            "siteLabel": "Ensembl Genome Browser",
-            "siteUri": "https://beta.ensembl.org"
-        }
-        output = json_format.MessageToJson(utils.create_release(input_data[0]))
-        assert json.loads(output) == expected_output
 
     def test_get_karyotype_information(self, genome_db_conn):
         output = json_format.MessageToJson(
@@ -316,339 +178,340 @@ class TestClass:
 
         expected_output = {
             "genomeUuid": "a73357ab-93e7-11ec-a39d-005056b38ce3",
-              "datasets": {
+            "datasets": {
                 "evidence": {
-                  "datasetInfos": [
-                    {
-                      "datasetUuid": "64a66f22-07a9-476e-9816-785e2ccb9c30",
-                      "datasetName": "evidence",
-                      "datasetLabel": "Manual Add",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "64a66f22-07a9-476e-9816-785e2ccb9c30",
-                      "datasetName": "evidence",
-                      "datasetLabel": "Manual Add",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "64a66f22-07a9-476e-9816-785e2ccb9c30",
-                      "datasetName": "evidence",
-                      "datasetLabel": "Manual Add",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "64a66f22-07a9-476e-9816-785e2ccb9c30",
-                      "datasetName": "evidence",
-                      "datasetLabel": "Manual Add",
-                      "version": 108.0
-                    }
-                  ]
+                    "datasetInfos": [
+                        {
+                            "datasetUuid": "64a66f22-07a9-476e-9816-785e2ccb9c30",
+                            "datasetName": "evidence",
+                            "datasetLabel": "Manual Add",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "64a66f22-07a9-476e-9816-785e2ccb9c30",
+                            "datasetName": "evidence",
+                            "datasetLabel": "Manual Add",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "64a66f22-07a9-476e-9816-785e2ccb9c30",
+                            "datasetName": "evidence",
+                            "datasetLabel": "Manual Add",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "64a66f22-07a9-476e-9816-785e2ccb9c30",
+                            "datasetName": "evidence",
+                            "datasetLabel": "Manual Add",
+                            "version": 108.0
+                        }
+                    ]
                 },
                 "assembly": {
-                  "datasetInfos": [
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
-                      "datasetName": "assembly",
-                      "datasetLabel": "GCA_900519105.1",
-                      "version": 108.0
-                    }
-                  ]
+                    "datasetInfos": [
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "b4ff55e3-d06a-4772-bb13-81c3207669e3",
+                            "datasetName": "assembly",
+                            "datasetLabel": "GCA_900519105.1",
+                            "version": 108.0
+                        }
+                    ]
                 },
                 "homologies": {
-                  "datasetInfos": [
-                    {
-                      "datasetUuid": "e67ca09d-2e7b-4135-a990-6a2d1bca7285",
-                      "datasetName": "homologies",
-                      "datasetLabel": "Manual Add",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "e67ca09d-2e7b-4135-a990-6a2d1bca7285",
-                      "datasetName": "homologies",
-                      "datasetLabel": "Manual Add",
-                      "version": 108.0
-                    }
-                  ]
+                    "datasetInfos": [
+                        {
+                            "datasetUuid": "e67ca09d-2e7b-4135-a990-6a2d1bca7285",
+                            "datasetName": "homologies",
+                            "datasetLabel": "Manual Add",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "e67ca09d-2e7b-4135-a990-6a2d1bca7285",
+                            "datasetName": "homologies",
+                            "datasetLabel": "Manual Add",
+                            "version": 108.0
+                        }
+                    ]
                 },
                 "genebuild": {
-                  "datasetInfos": [
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
-                      "datasetName": "genebuild",
-                      "datasetLabel": "2018-04-IWGSC",
-                      "version": 108.0
-                    }
-                  ]
+                    "datasetInfos": [
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "0dc05c6e-2910-4dbd-879a-719ba97d5824",
+                            "datasetName": "genebuild",
+                            "datasetLabel": "2018-04-IWGSC",
+                            "version": 108.0
+                        }
+                    ]
                 },
                 "variation": {
-                  "datasetInfos": [
-                    {
-                      "datasetUuid": "4d411e2d-676e-4fe0-b0d7-65a9e33fd47f",
-                      "datasetName": "variation",
-                      "datasetLabel": "Manual Add",
-                      "version": 108.0
-                    },
-                    {
-                      "datasetUuid": "4d411e2d-676e-4fe0-b0d7-65a9e33fd47f",
-                      "datasetName": "variation",
-                      "datasetLabel": "Manual Add",
-                      "version": 108.0
-                    }
-                  ]
+                    "datasetInfos": [
+                        {
+                            "datasetUuid": "4d411e2d-676e-4fe0-b0d7-65a9e33fd47f",
+                            "datasetName": "variation",
+                            "datasetLabel": "Manual Add",
+                            "version": 108.0
+                        },
+                        {
+                            "datasetUuid": "4d411e2d-676e-4fe0-b0d7-65a9e33fd47f",
+                            "datasetName": "variation",
+                            "datasetLabel": "Manual Add",
+                            "version": 108.0
+                        }
+                    ]
                 }
-              }
             }
+        }
         assert json.loads(output) == expected_output
 
     def test_get_datasets_list_by_uuid_no_results(self, genome_db_conn):
-        output = json_format.MessageToJson(utils.get_datasets_list_by_uuid(genome_db_conn, "some-random-uuid-f00-b4r", 103.0))
+        output = json_format.MessageToJson(
+            utils.get_datasets_list_by_uuid(genome_db_conn, "some-random-uuid-f00-b4r", 103.0))
         output = json.loads(output)
         expected_output = {}
         assert output == expected_output
@@ -656,7 +519,8 @@ class TestClass:
     def test_get_dataset_by_genome_and_dataset_type(self, genome_db_conn):
         # TODO: Fix
         output = json_format.MessageToJson(
-            utils.get_dataset_by_genome_and_dataset_type(genome_db_conn, "a7335667-93e7-11ec-a39d-005056b38ce3", "assembly")
+            utils.get_dataset_by_genome_and_dataset_type(genome_db_conn, "a7335667-93e7-11ec-a39d-005056b38ce3",
+                                                         "assembly")
         )
         output = json.loads(output)
         assert output == {'genomeUuid': 'a7335667-93e7-11ec-a39d-005056b38ce3', 'datasetType': 'assembly',
@@ -728,7 +592,8 @@ class TestClass:
 
     def test_get_dataset_by_genome_id_no_results(self, genome_db_conn):
         output = json_format.MessageToJson(
-            utils.get_dataset_by_genome_and_dataset_type(genome_db_conn, "a7335667-93e7-11ec-a39d-005056b38ce3", "blah blah blah"))
+            utils.get_dataset_by_genome_and_dataset_type(genome_db_conn, "a7335667-93e7-11ec-a39d-005056b38ce3",
+                                                         "blah blah blah"))
         output = json.loads(output)
         assert output == {}
 
@@ -745,7 +610,7 @@ class TestClass:
     def test_get_genome_by_uuid(self, genome_db_conn):
         output = json_format.MessageToJson(
             utils.get_genome_by_uuid(genome_db_conn,
-                                       "a73357ab-93e7-11ec-a39d-005056b38ce3", 108.0))
+                                     "a73357ab-93e7-11ec-a39d-005056b38ce3", 108.0))
         expected_output = {"assembly": {"accession": "GCA_900519105.1",
                                         "ensemblName": "IWGSC",
                                         "level": "chromosome",
@@ -902,7 +767,7 @@ class TestClass:
     def test_get_genomes_by_keyword_no_matches(self, genome_db_conn):
         output = list(
             utils.get_genomes_by_keyword_iterator(genome_db_conn, "bigfoot",
-                                                    1))
+                                                  1))
         assert output == []
 
     def test_get_genomes_by_name(self, genome_db_conn):
@@ -924,9 +789,9 @@ class TestClass:
                                        "releaseDate": "2023-05-15",
                                        "releaseLabel": "Beta Release 1",
                                        "releaseVersion": 108.0,
-                                        "siteLabel": "Ensembl Genome Browser",
-                                        "siteName": "Ensembl",
-                                        "siteUri": "https://beta.ensembl.org"},
+                                       "siteLabel": "Ensembl Genome Browser",
+                                       "siteName": "Ensembl",
+                                       "siteUri": "https://beta.ensembl.org"},
                            "taxon": {"scientificName": "Homo sapiens", "taxonomyId": 9606}}
         assert json.loads(output) == expected_output
 
@@ -948,8 +813,8 @@ class TestClass:
                                        "releaseDate": "2023-05-15",
                                        "releaseLabel": "Beta Release 1",
                                        "releaseVersion": 108.0,
-                                        "siteLabel": "Ensembl Genome Browser",
-                                        "siteName": "Ensembl",
-                                        "siteUri": "https://beta.ensembl.org"},
+                                       "siteLabel": "Ensembl Genome Browser",
+                                       "siteName": "Ensembl",
+                                       "siteUri": "https://beta.ensembl.org"},
                            "taxon": {"scientificName": "Homo sapiens", "taxonomyId": 9606}}
         assert json.loads(output) == expected_output
